refactor: crates, deps (#26)
* Refactor

* clean

* push

* cleanup

* notes

* fix

* fix blob

* wip, failing to generate valid execution

* fix: prover

* clean

* fix

* fix

* fix

* fix

* cargo machete

* overrwrite

* single block

* clean

* clean

* clean

* fix

* fix

* feat: add single block

* re-add run-client-native

* use

* refactor

* clean

* clean

---------
