+++ crates/client-utils/src/boot.rs
@@ -11,7 +11,7 @@ use serde::{Deserialize, Serialize};
 /// This struct contains all information needed to generate BootInfo,
 /// as the RollupConfig can be derived from the `chain_id`.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BootInfoWithoutRollupConfig {
+pub struct RawBootInfo {
     pub l1_head: B256,
     pub l2_output_root: B256,
     pub l2_claim: B256,
@@ -19,10 +19,10 @@ pub struct BootInfoWithoutRollupConfig {
     pub chain_id: u64,
 }
 
-impl From<BootInfoWithoutRollupConfig> for BootInfo {
+impl From<RawBootInfo> for BootInfo {
     /// Convert the BootInfoWithoutRollupConfig into BootInfo by deriving the RollupConfig.
-    fn from(boot_info_without_rollup_config: BootInfoWithoutRollupConfig) -> Self {
-        let BootInfoWithoutRollupConfig {
+    fn from(boot_info_without_rollup_config: RawBootInfo) -> Self {
+        let RawBootInfo {
             l1_head,
             l2_output_root,
             l2_claim,
@@ -42,29 +42,28 @@ impl From<BootInfoWithoutRollupConfig> for BootInfo {
     }
 }
 
-impl BootInfoWithoutRollupConfig {
+sol! {
+    struct RawBootInfoStruct {
+        bytes32 l1Head;
+        bytes32 l2PreRoot;
+        bytes32 l2PostRoot;
+        uint256 l2BlockNumber;
+        uint256 chainId;
+    }
+}
+
+impl RawBootInfo {
     /// ABI encode the boot info. This is used to commit to in the zkVM,
     /// so that we can verify on chain that the correct values were used in
     /// the proof.
     pub fn abi_encode(&self) -> Vec<u8> {
-        sol! {
-            struct PublicValuesStruct {
-                bytes32 l1Head;
-                bytes32 l2PreRoot;
-                bytes32 l2PostRoot;
-                uint256 l2BlockNumber;
-                uint256 chainId;
-            }
-        }
-
-        let public_values = PublicValuesStruct {
+        RawBootInfoStruct {
             l1Head: self.l1_head,
             l2PreRoot: self.l2_output_root,
             l2PostRoot: self.l2_claim,
             l2BlockNumber: U256::from(self.l2_claim_block),
             chainId: U256::from(self.chain_id),
-        };
-
-        public_values.abi_encode()
+        }
+        .abi_encode()
     }
 }
